requirements.txt
bt
pandas
yfinance
alpaca-py
matplotlib
PyYAML
numpy
seaborn
